+++ info.rkt
@@ -6,5 +6,5 @@
                      "gui-doc"))
 (define scribblings '(("scribblings/sauron.scrbl" (multi-page) (tool))))
 (define pkg-desc "A Racket IDE")
-(define version "0.2.1")
+(define version "0.3.0")
 (define pkg-authors '(dannypsnl))